Provato a inserire la classe Animator nel gioco ma ovviamente è buggato a causa della classe Graphics
