+++ src/Graphic/Frame.java
@@ -5,6 +5,8 @@ import java.awt.Rectangle;
 import javax.swing.JFrame;
 import javax.swing.JPanel;
 
+import GameObjects.Player;
+
 public class Frame extends JFrame{
 	
 	public static JPanel game = new Game();
@@ -19,6 +21,7 @@ public class Frame extends JFrame{
 		this.setBounds(new Rectangle(gameWidth, gameHeight));
 		this.setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
 		
+		((Game) game).setPlayer(new Player());
 		this.add(game);
 
 		this.setVisible(true);
